config-manager/requirements.txt
# Config Manager Dependencies
PyYAML>=6.0
Jinja2>=3.1.2
Flask>=2.3.0
Flask-CORS>=4.0.0